Implement HTML Generation.
Produce a single HTML file for each camera make, camera model and also an index.

The index HTML page must contain:
- Thumbnail images for the first 10 work
- Navigation that allows the user to browse to all camera makes

Each Camera Make HTML page must contain:
- Thumbnail images of the first 10 works for that camera make
- Navigation that allows the user to browse to the index page and to all camera models of that make

Each Camera Model HTML page must contain:
- Thumbnail images of all works for that camera make and model
- Navigation that allows the user to browse to the index page and the camera make

[Note]
I chose to sanitise the works collection by removing those work objects that lack either: thumbnail url /or/ Make name /or/ Model name. This decision was made based on common sense, because if you think about it, it wouldn't make sense to (for instance) display a work on the index page that is missing a model - user will keep drilling down model links to find out which model was used to take that particular picture, but won't be able to find the picture in any of the model pages. Of course I understand that, in a real world situation every functional detail must be clarified with the product owner/client, but due to our long question-response cycle time, I took the liberty to make the above assumption. :)  I hope it is alright with you.